+++ experimental/packages/opentelemetry-sdk-node/src/sdk.ts
@@ -152,9 +152,10 @@ export class NodeSDK {
     }
 
     if (configuration.logRecordProcessor) {
-      this._loggerProviderConfig = {
+      const loggerProviderConfig: LoggerProviderConfig = {
         logRecordProcessor: configuration.logRecordProcessor,
       };
+      this.configureLoggerProvider(loggerProviderConfig);
     }
 
     if (configuration.metricReader || configuration.views) {
@@ -197,6 +198,34 @@ export class NodeSDK {
       contextManager,
       textMapPropagator,
     };
+  }
+
+  /**
+   * @deprecated Please pass {@code logRecordProcessor} to the constructor options instead.
+   *
+   * Set configurations needed to register a LoggerProvider
+   */
+  public configureLoggerProvider(config: LoggerProviderConfig): void {
+    // nothing is set yet, we can set config and then return
+    if (this._loggerProviderConfig == null) {
+      this._loggerProviderConfig = config;
+      return;
+    }
+
+    // make sure we do not override existing logRecordProcessor with other logRecordProcessors.
+    if (
+      this._loggerProviderConfig.logRecordProcessor != null &&
+      config.logRecordProcessor != null
+    ) {
+      throw new Error(
+        'LogRecordProcessor passed but LogRecordProcessor has already been configured.'
+      );
+    }
+
+    // set logRecordProcessor, but make sure we do not override existing logRecordProcessors with null/undefined.
+    if (config.logRecordProcessor != null) {
+      this._loggerProviderConfig.logRecordProcessor = config.logRecordProcessor;
+    }
   }
 
   /**